+++ src/main/java/com/spectralink/API_SLK/model/entities/Product.java
@@ -13,13 +13,13 @@ import lombok.NoArgsConstructor;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false, length = 100)
     private String nombre;
 
     @Column(nullable = false)
-    private int precio;
+    private Integer precio;
 
     @Column(nullable = false)
     private String descripcion;
